Refactor Param inner classes to outer
