+++ src/main/java/dk/kb/discover/util/solrshield/SearchComponent.java
@@ -14,6 +14,7 @@
  */
 package dk.kb.discover.util.solrshield;
 
+import dk.kb.discover.util.solrshield.params.*;
 import dk.kb.util.yaml.YAML;
 
 /**
@@ -21,34 +22,34 @@ import dk.kb.util.yaml.YAML;
  */
 public class SearchComponent extends Component<SearchComponent> {
 
-    protected Param.StringParam q;
-    protected Param.StringParam fq;
-    protected Param.IntegerParam rows;
-    protected Param.IntegerParam start;
-    protected Param.FieldsParam fl;
-    protected Param.StringParam qOp;
-    protected Param.StringParam wt;
-    protected Param.StringParam version;
-    protected Param.BooleanParam indent;
-    protected Param.StringParam debug;
-    protected Param.BooleanParam debugExplainStructured;
+    protected StringParam q;
+    protected StringParam fq;
+    protected IntegerParam rows;
+    protected IntegerParam start;
+    protected FieldsParam fl;
+    protected StringParam qOp;
+    protected StringParam wt;
+    protected StringParam version;
+    protected BooleanParam indent;
+    protected StringParam debug;
+    protected BooleanParam debugExplainStructured;
 
     public SearchComponent(Profile profile, YAML config) {
         super(profile, "search", config);
 
         YAML paramsConf = config.getSubMap("params");
-        addParam(paramsConf, "q", paramConf -> this.q = new Param.StringParam(profile, paramConf, false));
-        addParam(paramsConf, "fq", paramConf -> this.fq = new Param.StringParam(profile, paramConf, true));
-        addParam(paramsConf, "rows", paramConf -> this.rows = new Param.IntegerParam(profile, paramConf));
-        addParam(paramsConf, "start", paramConf -> this.start = new Param.IntegerParam(profile, paramConf));
-        addParam(paramsConf, "fl", paramConf -> this.fl = new Param.FieldsParam(profile, paramConf));
+        addParam(paramsConf, "q", paramConf -> this.q = new StringParam(profile, paramConf, false));
+        addParam(paramsConf, "fq", paramConf -> this.fq = new StringParam(profile, paramConf, true));
+        addParam(paramsConf, "rows", paramConf -> this.rows = new IntegerParam(profile, paramConf));
+        addParam(paramsConf, "start", paramConf -> this.start = new IntegerParam(profile, paramConf));
+        addParam(paramsConf, "fl", paramConf -> this.fl = new FieldsParam(profile, paramConf));
         // TODO: Add support for enums
-        addParam(paramsConf, "q.op", paramConf -> this.qOp = new Param.StringParam(profile, paramConf, false));
-        addParam(paramsConf, "wt", paramConf -> this.wt = new Param.StringParam(profile, paramConf, false));
-        addParam(paramsConf, "version", paramConf -> this.version = new Param.StringParam(profile, paramConf, false));
-        addParam(paramsConf, "indent", paramConf -> this.indent = new Param.BooleanParam(profile, paramConf));
-        addParam(paramsConf, "debug", paramConf -> this.debug = new Param.StringParam(profile, paramConf, false));
-        addParam(paramsConf, "debug.explain.structured", paramConf -> this.debugExplainStructured = new Param.BooleanParam(profile, paramConf));
+        addParam(paramsConf, "q.op", paramConf -> this.qOp = new StringParam(profile, paramConf, false));
+        addParam(paramsConf, "wt", paramConf -> this.wt = new StringParam(profile, paramConf, false));
+        addParam(paramsConf, "version", paramConf -> this.version = new StringParam(profile, paramConf, false));
+        addParam(paramsConf, "indent", paramConf -> this.indent = new BooleanParam(profile, paramConf));
+        addParam(paramsConf, "debug", paramConf -> this.debug = new StringParam(profile, paramConf, false));
+        addParam(paramsConf, "debug.explain.structured", paramConf -> this.debugExplainStructured = new BooleanParam(profile, paramConf));
     }
 
     @Override
